Add strict and offline index support (#36)
Creates two separate indexes, `packages/local` and `packages/all`
instead of just `packages/all`. `packages/all` inherits from `root/pypi`
and `packages/local`. `packages/local` does not allow packages to be
downloaded from PyPI.

All of the "build" dependencies must be published to `packages/local`
for it to be used to install source distributions. This pull request
vendors all of the needed dependencies.

Additionally, there is an `--offline` flag for `packse index up` which
does not allow any PyPI access. This is kind of nice when you are not
planning on using fallback to PyPI as it prevents the devpi server from
doing any additional indexing of the real PyPI.

This was tested by installing a scenario published to the strict and
offline indexes.