add a collection of tests that use local versions (#132)
This improves our coverage for installing packages involving versions
with local segments.

~~This PR also adds a new optional `expected.ignore` property. When it's
true, it's expected that the scenario will be ignored in downstream
tools. For example, if the scenario doesn't pass yet even though it is
expected or desired to.~~

I've backed out the change to add `expected.ignore` to the schema, and
instead opted to mark tests as skipped on `uv`'s end in
https://github.com/astral-sh/uv/pull/2022.
This makes sense to me since this de-couples the current state of `uv`
from
packse scenarios.