templates/package: fix template to use triple quotes
I noticed that when using a requirement like

```
c<2.0.0 ; sys_platform == 'linux'
```

then the single quotes seem to get rewritten as double quotes. And since
the TOML template just prints the value as-is without any escaping, it
ends up producing an invalid TOML file.

I do wonder whether a better fix here would be to abdicate templates and
use a proper TOML serializer from an in-memory data structure, but this
fix is probably good enough for now. Namely, we use triply quoted TOML
strings. The only way this can go wrong is if the requirement contains
three single quotes in succession.