scenario: add resolve_options.universal option
In order for us to test a universal resolver, we need to break a number
of assumptions made by `pip compile`-esque tools. Firstly, the tool
itself needs to be aware of it. (In the case of `uv`, it's a matter of
running `uv pip compile` versus `uv lock`.) Secondly, the "expected"
packages could wind up with more than one version for a single package.