Add wrong-backtracking scenario (#230)
Scenario for https://github.com/astral-sh/uv/issues/8157.

There are two packages, `a` and `b`. We select `a` with `a==2.0.0`
first, and then `b`, but `a==2.0.0` conflicts with all new versions of
`b`, so we backtrack through versions of `b`.

We need to detect this conflict and prioritize `b` over `a` instead of
backtracking down to the too old version of `b==1.0.0` that doesn't
depend on `a` anymore.